Create Agent.pl
This is the logic which controls my agent to navigate the wumpus world maze. It will always start each time by checking to see if it can pick up gold, and then try to move around based upon the current perception of the world. The agent will also keep track of its position while moving around so that it can attempt to leave the maze upon finding gold.